% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc.theta}
\alias{calc.theta}
\title{Calculate covariance parameter vector (theta)}
\usage{
calc.theta(VarCov, sigma)
}
\arguments{
\item{VarCov}{variance-covariance components of random effects.
If there are multiple random effects of one grouping factor, provide the variance-covariance matrix.
If there are multiple grouping factors, supply the variance-covariance matrix of each grouping factor in a list.}

\item{sigma}{error standard deviation.}
}
\value{
covariance parameter vector
}
\description{
This function calculates the covariance parameters
}
\details{
For more details on the structure and estimation of \code{theta}, refer to the documentation
of the \code{\link[lme4]{getME}} function in the \code{lme4} package.
}
\seealso{
\code{\link[lme4]{getME}}, \code{\link[lme4]{lmer}}
}
