% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_calculation.R
\name{contestMD}
\alias{contestMD}
\title{Computes power of F-test for multi-dimensional contrast matrices}
\usage{
contestMD(object, L, sig.level = 0.05, eps = sqrt(.Machine$double.eps))
}
\arguments{
\item{object}{design object}

\item{L}{contrast matrix}

\item{sig.level}{significance level}

\item{eps}{numeric tolerance}
}
\value{
A data frame
}
\description{
Computes power of F-test for multi-dimensional contrast matrices
}
\keyword{internal}
