% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corClasses.R
\name{corCAR1}
\alias{corCAR1}
\title{Continuous AR(1) Correlation Structure}
\usage{
corCAR1(value, form, fixed)
}
\arguments{
\item{value}{numeric value for the correlation parameter}

\item{form}{a one-sided formula of the form \code{~ t}, \code{~ 1 | g}, or \code{~ t | g},
specifying a time covariate \code{t}, a grouping factor \code{g}, or both.}

\item{fixed}{unused}
}
\description{
Re-exports \link[nlme:corCAR1]{nlme::corCAR1} from the \pkg{nlme} package.
}
\seealso{
\link{corClasses}.
See \link[nlme:corCAR1]{nlme::corCAR1} for original documentation.
}
