% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corClasses.R
\name{corCompSymm}
\alias{corCompSymm}
\title{Compound Symmetry Correlation Structure}
\usage{
corCompSymm(value, form, fixed)
}
\arguments{
\item{value}{numeric vector of parameter values}

\item{form}{a one-sided formula of the form \code{~ 1 | g} specifying a grouping factor.}

\item{fixed}{unused}
}
\description{
Re-exports \link[nlme:corCompSymm]{nlme::corCompSymm} from the \pkg{nlme} package.
}
\seealso{
\link{corClasses}
See \link[nlme:corCompSymm]{nlme::corCompSymm} for original documentation.
See \link[nlme:corCompSymm]{nlme::corCompSymm} for original documentation.
}
