% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corClasses.R
\name{corSpher}
\alias{corSpher}
\title{Spherical Correlation Structure}
\usage{
corSpher(value, form, nugget, metric, fixed)
}
\arguments{
\item{value}{numeric value(s) for the parameter(s)}

\item{form}{a one-sided formula of the form \code{~ S1 + ... + Sp}, or \code{~ S1 + ... + Sp | g}
spatial covariates S1 through Sp and, optionally, a grouping factor g.}

\item{nugget}{logical; if \code{TRUE} a nugget effect is added}

\item{metric}{character string specifying the distance metric}

\item{fixed}{unused}
}
\description{
Re-exports \link[nlme:corSpher]{nlme::corSpher} from the \pkg{nlme} package.
}
\seealso{
\link{corClasses}
See \link[nlme:corSpher]{nlme::corSpher} for original documentation.
}
