% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corClasses.R
\name{corSymm}
\alias{corSymm}
\title{General Correlation Structure}
\usage{
corSymm(value, form, fixed)
}
\arguments{
\item{value}{numeric vector of correlation parameter values}

\item{form}{a one-sided formula of the form \code{~ t}, \code{~ 1 | g}, or \code{~ t | g},
specifying a covariate \code{t}, a grouping factor \code{g}, or both. A covariate indicates
the order of the data rows in the residual error matrix of each group.}

\item{fixed}{unused}
}
\description{
Re-exports \link[nlme:corSymm]{nlme::corSymm} from the \pkg{nlme} package.
}
\seealso{
\link{corClasses}
See \link[nlme:corSymm]{nlme::corSymm} for original documentation.
}
