% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-generating.R
\name{designCRD}
\alias{designCRD}
\alias{design.CRD}
\alias{design.RCBD}
\alias{design.LSD}
\alias{design.COD}
\alias{design.SPD}
\alias{designRCBD}
\alias{designLSD}
\alias{designCOD}
\alias{designSPD}
\title{Creation of Experimental Designs}
\usage{
designCRD(
  treatments,
  label,
  replicates,
  formula,
  beta = NULL,
  means = NULL,
  sigma2,
  template = FALSE
)

designRCBD(
  treatments,
  label,
  blocks,
  formula,
  beta = NULL,
  means = NULL,
  vcomp,
  sigma2,
  template = FALSE
)

designLSD(
  treatments,
  label,
  squares = 1,
  reuse = c("row", "col", "both"),
  formula,
  beta = NULL,
  means = NULL,
  vcomp,
  sigma2,
  template = FALSE
)

designCOD(
  treatments,
  label,
  squares = 1,
  formula,
  beta = NULL,
  means = NULL,
  vcomp,
  sigma2,
  template = FALSE
)

designSPD(
  trt.main,
  trt.sub,
  label,
  replicates,
  formula,
  beta = NULL,
  means = NULL,
  vcomp,
  sigma2,
  template = FALSE
)
}
\arguments{
\item{treatments}{An integer-valued vector specifying the treatment structure,
in which the length of the vector indicates the number of treatment factors,
and each value represents the number of levels for each factor. A maximum of
two factors is allowed, and they are arranged in a factorial design.
For instance, \code{treatments = n} specifies one treatment factor with n
levels, and \code{treatments=c(2,3)} creates a "2x3" factorial design of
two treatment factors with 2 and 3 levels, respectively.}

\item{label}{A optional list of strings specifying the names of
treatment factors and factor levels. Each vector in the list represents a
treatment factor, where the name of the vector specifies the name of the
factor, and the values in the vector are the labels for that factor's levels.
If not provided, factors and levels for one and two treatment factors are
labeled as \code{list(trt = c("1", "2", ...))} and
\code{list(facA = c("1", "2", ...), facB = c("1", "2", ...))}, respectively.}

\item{replicates}{The number of experimental units per treatment in a completely
randomized design or the number of experimental units (main plots) per treatment
of main plot factors.}

\item{formula}{a right-hand-side formula to be used for testing treatment effects
Use the syntax of \code{\link{lm}} for fixed effects and \link[lme4]{lmer} for random effects.
If not specified, a default formula with main effects and all interactions is used internally.}

\item{beta}{model coefficients, an optional input of fixed effects.
The coefficients of factor variables are the coefficients of dummy variables created
using "contr.treatment".}

\item{means}{marginal means or conditioned means if factors have interaction.
Regression coefficients should be provided for numerical variables. Use \code{template = TRUE} to create
a template for these inputs.}

\item{sigma2}{error variance.}

\item{template}{if TRUE, a template for \code{beta}, \code{means}, and \code{vcomp} is
generated to indicate the orders of inputs}

\item{blocks}{The number of blocks.}

\item{vcomp}{variance-covariance parameters of random effects provided in a vector.
Use \code{template = TRUE} to create a template for these inputs.}

\item{squares}{The number of replicated squares. By default, 1, i.e., no
replicated squares.}

\item{reuse}{A character string specifying how to replicate squares when
there are multiple squares. Options are: "row" for reusing row blocks, "col"
for reusing column blocks, or "both" for reusing both row and column blocks
to replicate a single square.}

\item{trt.main}{An integer-valued vector specifying the treatment structure at
main plot level for a split plot design, similar to \code{treatments}.}

\item{trt.sub}{An integer-valued vector specifying the treatment structure at
sub plot level for a split plot design, similar to \code{treatments}.}
}
\value{
a list with critical components for power analysis
}
\description{
These functions are used to create some common designs in agricultural studies by specifying the design characteristics.
}
\details{
Each function creates a specific design as described below:
\describe{
\item{\code{designCRD}}{Completely Randomized Design.
By default, the model formula is \code{y ~ trt} for one factor and
\code{y ~ facA*facB} for two factors, unless explicitly specified. If the
\code{label} argument is provided, the formula is automatically updated with the
specified treatment factor names.}
\item{\code{designRCBD}}{Randomized Complete Block Design.
The default model formula is \code{y ~ trt + (1|block)} for one factor and
\code{y ~ facA*facB + (1|block)} for two factors. If \code{label} is provided, the
fixed effect parts of the formula are automatically updated with the specified
names. The label of block factor ("block") in the formula is not changeable.}
\item{\code{designLSD}}{Latin Square Design.
The default formula is \code{y ~ trt + (1|row) + (1|col)} for one factor and
\code{y ~ facA*facB + (1|row) + (1|col)} for two factors. If \code{label} is provided,
the fixed effect parts of the formula are automatically updated with the specified
names. The labels of row ("row") and column ("col") block factors are not changeable.}
\item{\code{designCOD}}{Crossover Design, which is a special case of LSD
with time periods and individuals as blocks. Period blocks are reused when
replicating squares.
The default formula is \code{y ~ trt + (1|subject) + (1|period)} for one factor
and \code{y ~ facA*facB + (1|subject) + (1|period)} for two factors. If \code{label}
is provided, the fixed effect parts of the formula are automatically updated
with the specified names. Note that "subject" and "period" are the labels for
the two blocking factors and cannot be changed.}
\item{\code{designSPD}}{Split Plot Design.
The default formula includes the main effects of all treatment factors at
both the main and sub-plot levels, their interactions, and the random effects
of main plots: \code{y ~ . + (1|mainplot)}. If \code{label} is provided, the fixed
effect parts of the formula are automatically updated with the specified names.
The experimental unit at the main plot level (i.e., the block factor at the
subplot level) is always denoted as "mainplot".}
\item{\code{designCustom}}{Customized Design.}
}
}
\examples{
# Example 1: Evaluate the power of a CRD with one treatment factor

## Create a design object

crd <- designCRD(
  treatments = 4, # 4 levels of one treatment factor
  replicates = 12, # 12 units per level, 48 units totally
  means = c(30, 28, 33, 35), # means of the 4 levels
  sigma2 = 10 # error variance
)

## power of omnibus test
pwr.anova(crd)

## power of contrast
pwr.contrast(crd, which = "trt", contrast = "pairwise") # pairwise comparisons
pwr.contrast(crd, which = "trt", contrast = "poly") # polynomial contrasts

# More examples are available in the package vignette("pwr4exp")
# and on the package website: https://an-ethz.github.io/pwr4exp/

}
\seealso{
\link{mkdesign}, \code{\link[=pwr.anova]{pwr.anova()}}, \code{\link[=pwr.contrast]{pwr.contrast()}}
}
