% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{class}
\name{customLmerMod-class}
\alias{customLmerMod-class}
\title{Extend lmerModLmerTest class}
\description{
This class extends \code{lmerModLmerTest} by adding a \code{DenDF} slot.
}
\section{Slots}{

\describe{
\item{\code{DenDF}}{Numeric vector of denominator degrees of freedom.}
}}

