% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-generating.R
\name{designCOD}
\alias{designCOD}
\title{Create a Crossover Design}
\usage{
designCOD(treatments, label, squares, formula, beta, VarCov, sigma2, ...)
}
\arguments{
\item{treatments}{a vector specifying the treatment structure, in which the
length of the vector indicates the number of treatment factors, and each value
represents the number of levels for each factor. For example, \code{treatments=c(2,3)}
specifies two treatment factors with 2 and 3 levels, respectively.}

\item{label}{Optional. A list specifying the names of treatment factors and
their corresponding level labels. Each element of the list represents a factor,
with the name of the element being the name of the factor. The value of each
element should be a vector containing the labels for that factor's levels.
By default, the name of one treatment factor is "trt," and the names of two
treatment factors are "facA" and "facB," respectively.
The levels of each factor are coded as "1", "2", ..., if not specified in \code{label}.
For example, \code{list(trt = c("A1", "A2"))} assigns the name "trt" to the treatment factor,
with its levels labeled as "A1" and "A2." For multiple factors, such as
\code{list(A = c("A1", "A2"), B = c("B1", "B2", "B3"))}, the first factor is named
"A" with levels "A1" and "A2," and the second factor is named "B" with levels
"B1," "B2," and "B3."}

\item{squares}{the number of replicated squares}

\item{formula}{a symbolic description of the model that would be used to test
the treatment effects in post-experimental data analysis.
The way of writing formula , see \code{\link{lm}} for fixed effects and
\code{\link{lmer}} for random effects.
The default formula for one treatment factor and two treatment factors are
\code{y ~ trt + (1|subject) + (1|period)} and \code{y ~ facA*facB + (1|subject) + (1|period)}, respectively,
and when the \code{label} argument is supplied, the names of treatment factors are
replaced by the corresponding names specified in \code{label}.
The response variable in the formula is always denoted as \code{y}, and row and column block factors
are always denoted as \code{subject} and \code{period}.}

\item{beta}{a vector of the expectations of model coefficients.The coefficients
for categorical variables are the coefficients of dummy variables created using
\code{\link{contr.treatment}} contrast coding.}

\item{VarCov}{variance-covariance components of random effects.
If there are multiple random effects of one grouping factor, provide the variance-covariance matrix.
If there are multiple grouping factors, supply the variance-covariance matrix of each grouping factor in a list.}

\item{sigma2}{error variance}

\item{...}{other arguments passed to the \code{anova} function in \code{lmerTest}.
The type of ANOVA table, with Type III as the default, and the method for computing the denominator degrees of freedom, with Satterthwaite's method as the default, can be changed.
For more details, see \link[lmerTest]{anova.lmerModLmerTest}.}
}
\value{
an object containing design name, design data frame, model formula,
and a pseudo model object with the expected effect size and variance-covariance
components of random effects.
}
\description{
A crossover design is a specific type case of LSD, where each time period serves as a block,
and period blocks are reused during replicating squares.
}
