% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-generating.R
\name{designCRD}
\alias{designCRD}
\title{Create a Completely Randomized Design}
\usage{
designCRD(treatments, label, replicates, formula, beta, sigma2)
}
\arguments{
\item{treatments}{a vector specifying the treatment structure,
in which the length of the vector indicates the number of treatment factors,
and each value represents the number of levels for each factor.
A maximum of two factors is allowed, and they are arranged in a factorial design.
For example, \code{treatments=c(2,3)} specifies a "2x3" factorial design of
two treatment factors with 2 and 3 levels, respectively.}

\item{label}{Optional. A list specifying the names of treatment factors and
their corresponding level labels. Each element of the list represents a factor,
with the name of the element being the name of the factor. The value of each
element should be a vector containing the labels for that factor's levels.
By default, the name of one treatment factor is "trt," and the names of two
treatment factors are "facA" and "facB," respectively.
The levels of each factor are coded as "1", "2", ..., if not specified in \code{label}.
For example, \code{list(trt = c("A1", "A2"))} assigns the name "trt" to the treatment factor,
with its levels labeled as "A1" and "A2." For multiple factors, such as
\code{list(A = c("A1", "A2"), B = c("B1", "B2", "B3"))}, the first factor is named
"A" with levels "A1" and "A2," and the second factor is named "B" with levels
"B1," "B2," and "B3."}

\item{replicates}{The number of experimental units per group.
In a one-factor design, 'group' refers to the levels of that factor.
In a two-factor design, 'group' refers to each combination of levels between the two factors.}

\item{formula}{a symbolic description of the model that would be used to test
the treatment effects in post-experimental data analysis.
The way of writing formula , see \code{\link{lm}} for fixed effects and
\code{\link{lmer}} for random effects.
The default formula for one treatment factor and two treatment factors are
\code{y ~ trt } and \code{y ~ facA*facB}, respectively,
and when the \code{label} argument is supplied, the names of treatment factors are
replaced by the corresponding names specified in \code{label}.
The response variable in the formula is always denoted as \code{y}.}

\item{beta}{a vector of the expectations of model coefficients.The coefficients
for categorical variables are the coefficients of dummy variables created using
\code{\link{contr.treatment}} contrast coding.}

\item{sigma2}{error variance}
}
\value{
an object containing design name, design data frame, model formula,
and a pseudo model object with the expected effect size and variance-covariance
components of random effects.
}
\description{
Create a Completely Randomized Design
}
