% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-generating.R
\name{designCustom}
\alias{designCustom}
\title{Create a Customized Design}
\usage{
designCustom(design.df, formula, beta, VarCov, sigma2, design.name, ...)
}
\arguments{
\item{design.df}{a data frame containing values of all independent variables as columns}

\item{formula}{an object of class \code{\link{formula}}: a symbolic description
of the model that would be used to test effects in post-experimental data analysis.
For details on model specification, see \code{\link{lm}} for fixed effects and
\code{\link{lmer}} for random effects.}

\item{beta}{a vector of the expectations of model coefficients.The coefficients
for categorical variables are the coefficients of dummy variables created using
\code{\link{contr.treatment}} contrast coding.}

\item{VarCov}{variance-covariance components of random effects.
If there are multiple random effects of one grouping factor, provide the variance-covariance matrix.
If there are multiple grouping factors, supply the variance-covariance matrix of each grouping factor in a list.}

\item{sigma2}{error variance}

\item{design.name}{optional, name of the design}

\item{...}{other arguments passed to the \code{anova} function in \code{lmerTest}.
The type of ANOVA table, with Type III as the default, and the method for computing the denominator degrees of freedom, with Satterthwaite's method as the default, can be changed.
For more details, see \link[lmerTest]{anova.lmerModLmerTest}.}
}
\value{
an object containing design name, design data frame, model formula,
and a pseudo model object with the expected effect size and variance-covariance
components of random effects.
}
\description{
Create a Customized Design
}
