% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-generating.R
\name{designSPD}
\alias{designSPD}
\title{Create a Split Plot Design}
\usage{
designSPD(
  trt.main,
  trt.sub,
  label,
  replicates,
  formula,
  beta,
  VarCov,
  sigma2,
  ...
)
}
\arguments{
\item{trt.main}{a vector specifying the treatment structure at main plot level, in which the
length of the vector indicates the number of main plot factors, and each value
represents the number of levels for each factor.}

\item{trt.sub}{a vector specifying the treatment structure at main plot level, in which the
length of the vector indicates the number of main plot factors, and each value
represents the number of levels for each factor.}

\item{label}{Optional. A list specifying the names of treatment factors and
their corresponding level labels. Each element of the list represents a factor,
with the name of the element being the name of the factor. The value of each
element should be a vector containing the labels for that factor's levels.
By default, the name of one treatment factor is "trt," and the names of two
treatment factors are "facA" and "facB," respectively.
The levels of each factor are coded as "1", "2", ..., if not specified in \code{label}.
For example, \code{list(trt = c("A1", "A2"))} assigns the name "trt" to the treatment factor,
with its levels labeled as "A1" and "A2." For multiple factors, such as
\code{list(A = c("A1", "A2"), B = c("B1", "B2", "B3"))}, the first factor is named
"A" with levels "A1" and "A2," and the second factor is named "B" with levels
"B1," "B2," and "B3."}

\item{replicates}{the number of main plots per treatment group of main plot factors}

\item{formula}{a symbolic description of the model that would be used to test
the treatment effects in post-experimental data analysis.
The way of writing formula , see \code{\link{lm}} for fixed effects and
\code{\link{lmer}} for random effects.
The response variable in the formula is always denoted as \code{y}, and the experimental
unit at main plot level (aka. block factor at subplot level) is always denoted as \code{mainplots}.
The default formula contains main effects of all treatment factors at main and sub- plot levels,
interactions between them factors, and random effects of main plots \code{y ~ . + (1|mainplots)}.
replaced by the corresponding names specified in \code{label}.
Column factor \code{(1|col)} in the default formula is replaced by \code{(1|square/col)} when
row factors are reused, and row factor \code{(1|row)} is replaced by \code{(1|square/row)} when
column factors are reused. When both row and column factors are resued, the random
effects in the model are denoted as \code{(1|square) + (1|square:row) + (1|square:col)}.}

\item{beta}{a vector of the expectations of model coefficients.The coefficients
for categorical variables are the coefficients of dummy variables created using
\code{\link{contr.treatment}} contrast coding.}

\item{VarCov}{variance-covariance components of random effects.
If there are multiple random effects of one grouping factor, provide the variance-covariance matrix.
If there are multiple grouping factors, supply the variance-covariance matrix of each grouping factor in a list.}

\item{sigma2}{error variance}

\item{...}{other arguments passed to the \code{anova} function in \code{lmerTest}.
The type of ANOVA table, with Type III as the default,
and the method for computing the denominator degrees of freedom,
with Satterthwaite's method as the default, can be changed.
For more details, see \link[lmerTest]{anova.lmerModLmerTest}.}
}
\value{
an object containing design name, design data frame, model formula,
and a pseudo model object with the expected effect size and variance-covariance
components of random effects.
}
\description{
Create a Split Plot Design
}
