% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_calculator.R
\name{designSplitplot}
\alias{designSplitplot}
\title{Create Split Plot Design}
\usage{
designSplitplot(
  trt.main,
  trt.sub,
  replications,
  formula,
  beta,
  VarCov,
  sigma2,
  ...
)
}
\arguments{
\item{trt.main}{a vector representing the number of levels of main plot factors sequentially.}

\item{trt.sub}{a vector representing the number of levels of subplot factors sequentially.}

\item{replications}{number of main plots in each main plot group}

\item{formula}{an object of class \code{\link{formula}}: a symbolic description of the model used to assess treatment effects, which should contain fixed components and sources of variation according to experimental designs, assumptions, and objectives. The details of model specification for fixed effects see \code{\link{lm}} and random effects see \code{\link{lmer}}.}

\item{beta}{effect size: a vector including the expectations of intercept, main effects of factors and interactions between factors. Vector values are in the same order as the coefficients of the model that would be fitted using the exact data with \code{\link{contr.treatment}}.}

\item{VarCov}{variance-covariance components of random effects. If there are multiple random effects of one grouping factor, supply variance-covariance components in a matrix with the same order as the random effects specified in model formula. If there are multiple grouping factors, supply variance-covariance matrix for each grouping factor in a list.}

\item{sigma2}{error variance}

\item{...}{additional arguments to be passed to \code{fit.pseu.model}}
}
\value{
a list object containing design name, design data frame, model formula, and a pseudo model object with the expected effect size and variance-covariance components of random effects.
}
\description{
Create Split Plot Design
}
\examples{
mydesign <- designSplitplot(trt.main = 2,
                            trt.sub = 4,
                           replications = 6,
                           beta = c(10, -2, 3, 4, 5, 1, 3, 0),
                           VarCov = 4,
                           sigma2 = 6)
}
