% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_design.R
\name{df.crossover}
\alias{df.crossover}
\title{Create data frame for Crossover design}
\usage{
df.crossover(treatments, squares)
}
\arguments{
\item{treatments}{a vector representing the number of levels of treatment factors sequentially. For example, \code{treatments=c(2,3)} specifies two treatment factors with 2 and 3 levels, respectively.}

\item{squares}{number of replicated squares}
}
\value{
a data.frame with columns for treatment factors, individuals (row block factor), period (column block factor), and squares
}
\description{
Create data frame for Crossover design
}
\examples{
df.crossover(4, 5)
}
