% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_design.R
\name{df.lsd}
\alias{df.lsd}
\title{Create data frame for Latin square design}
\usage{
df.lsd(treatments, squares = 1, reuse = c("row", "col", "both"))
}
\arguments{
\item{treatments}{a vector representing the number of levels of treatment factors sequentially. For example, \code{treatments=c(2,3)} specifies two treatment factors with 2 and 3 levels, respectively.}

\item{squares}{number of replicated squares}

\item{reuse}{a character string: "row", "col", or "both", indicating reuse of rows or columns or both when replicate a Latin square}
}
\value{
a data.frame with columns for treatment factors, row and column block factors, and squares
}
\description{
Create data frame for Latin square design
}
\examples{
df.lsd(5, 3)
}
