% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_design.R
\name{df.rcb}
\alias{df.rcb}
\title{Create data frame for randomized complete block design}
\usage{
df.rcb(treatments, blocks)
}
\arguments{
\item{treatments}{a vector representing the number of levels of treatment factors sequentially. For example, \code{treatments=c(2,3)} specifies two treatment factors with 2 and 3 levels, respectively.}

\item{blocks}{number of blocks}
}
\value{
a data.frame with columns for blocks and treatment factors
}
\description{
Create data frame for randomized complete block design
}
\examples{
df.rcb(5, 2)
}
