% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df.R
\name{df.spd}
\alias{df.spd}
\title{Create data frame for split-plot design}
\usage{
df.spd(trt.main, trt.sub, label, replicates)
}
\arguments{
\item{trt.main}{a vector specifying the treatment structure at main plot level, in which the
length of the vector indicates the number of main plot factors, and each value
represents the number of levels for each factor.}

\item{trt.sub}{a vector specifying the treatment structure at main plot level, in which the
length of the vector indicates the number of main plot factors, and each value
represents the number of levels for each factor.}

\item{label}{optional, names of treatments factors and their levels can be specified in a list.}

\item{replicates}{the number of main plots per treatment group of main plot factors}
}
\value{
a data.frame with columns for main plots, main treatments, and sub-treatments
}
\description{
Create data frame for split-plot design
}
