% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_design.R
\name{df.splitplot}
\alias{df.splitplot}
\title{Create data frame for split-plot design}
\usage{
df.splitplot(trt.main, trt.sub, replications)
}
\arguments{
\item{trt.main}{a vector representing the number of levels of main plot treatment factors sequentially.}

\item{trt.sub}{a vector representing the number of levels of sub plot treatment factors sequentially.}

\item{replications}{number of replications at main plots}
}
\value{
a data.frame with columns for main plots, main treatments, and sub-treatments
}
\description{
Create data frame for split-plot design
}
\examples{
df.splitplot(3, 4, 15)
}
