% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expandDoubleVerts}
\alias{expandDoubleVerts}
\title{Expand terms with \code{'||'} notation into separate \code{'|'} terms}
\usage{
expandDoubleVerts(term)
}
\arguments{
\item{term}{a formula}
}
\value{
the modified term
}
\description{
Expand terms with \code{'||'} notation into separate \code{'|'} terms
}
\keyword{functions}
\keyword{internal}
\keyword{lme4}
