% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factorize}
\alias{factorize}
\title{Convert variables to factors as necessary}
\usage{
factorize(x, frloc, char.only = FALSE)
}
\arguments{
\item{x}{a formula}

\item{frloc}{a data frame}

\item{char.only}{(logical) convert only character variables to factors?}
}
\value{
a copy of the data frame with factors converted
}
\description{
Convert variables to factors as necessary
}
\keyword{functions}
\keyword{internal}
\keyword{lme4}
