% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findbars}
\alias{findbars}
\title{Determine random-effects expressions from a formula}
\usage{
findbars(term)
}
\arguments{
\item{term}{a mixed-model formula}
}
\value{
pairs of expressions that were separated by vertical bars
}
\description{
Determine random-effects expressions from a formula
}
\section{Note}{
 This function is called recursively on individual
terms in the model, which is why the argument is called \code{term} and not
a name like \code{form}, indicating a formula.
}

\keyword{functions}
\keyword{internal}
\keyword{lme4}
