% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{milk}
\alias{milk}
\title{An exemplary dataset of a 4x4 crossover design with 2 squares}
\format{
A data frame with 32 rows and 4 variables:
\describe{
\item{Cow}{Factor: Cow index (8 levels)}
\item{Period}{Factor: Period index (4 levels)}
\item{Treatment}{Factor: Treatment index (4 levels)}
\item{MilkYield}{Numeric: milk yield recordings (in kg)}
}
}
\source{
Simulated data for package demonstration purposes.
}
\usage{
milk
}
\description{
Milk yield records from 8 cows over 4 different periods in a 4x4 crossover design.
The design includes 2 Latin squares, each consisting of 4 cows and 4 periods.
}
\keyword{datasets}
