% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdesign.R
\name{mkdesign}
\alias{mkdesign}
\title{Create any design}
\usage{
mkdesign(
  formula,
  data,
  beta = NULL,
  means = NULL,
  vcomp = NULL,
  sigma2 = NULL,
  correlation = NULL,
  REML = TRUE,
  template = FALSE,
  ...
)
}
\arguments{
\item{formula}{a right-hand-side formula to be used for testing treatment effects.
This formula follows the same syntax used in the lme4 package to specify fixed and random effects.}

\item{data}{a data frame containing all independent variables specified in the model,
which mirrors the structure of the data generated by the design}

\item{beta}{a vector of model coefficients.
The coefficients of factor variables are the coefficients of dummy variables created using "contr.treatment".
This is an optional input of fixed effects.}

\item{means}{a vector of marginal means or conditioned means if factors have interaction.
Regression coefficients are required for numerical variables. Either \code{beta} or \code{means} is required as input.
A template can be created when \code{template = TRUE}, or when only the formula and data are provided.
This will display the model coefficients or means, along with their respective order in the input, as needed.}

\item{vcomp}{a vector of variance-covariance components for random effects.
A template can also be created when \code{template = TRUE}, or when only the formula and data are provided.
The template represents a variance-covariance matrix, with the values indicating the order of the corresponding variance-covariance parameters that is required in vcomp.}

\item{sigma2}{error (residual) variance}

\item{correlation}{correlation structures, \code{corClasses} in \code{nlme}.}

\item{REML}{default, TRUE}

\item{template}{if TRUE, a template for \code{beta}, \code{means}, and \code{vcomp} is
generated to indicate the orders of inputs}

\item{...}{other arguments passed to internal functions}
}
\value{
a list with critical components of the design needed for power calculation
}
\description{
A design can be created if a data frame, a model formula, and other parameters are provided.
}
\examples{
# create a data frame,
## representing a CRD with repeated measures at 8 time points

n_subject = 6
n_trt = 3
n_hour = 8
trt = c("CON", "TRT1", "TRT2")

df.rep <- data.frame(
  subject = as.factor(rep(seq_len(n_trt*n_subject), each = n_hour)),
  hour = as.factor(rep(seq_len(n_hour), n_subject*n_trt)),
  trt = rep(trt, each = n_subject*n_hour)
)

# templates
temp <- mkdesign(formula = ~ trt*hour, data = df.rep)
temp$fixeff
temp$vcov

# create the design

design.rep <- mkdesign(
  formula = ~ trt*hour,
  data = df.rep,
  means =  c(1, 2.50, 3.5,
             1, 3.50, 4.54,
             1, 3.98, 5.80,
             1, 4.03, 5.4,
             1, 3.68, 5.49,
             1, 3.35, 4.71,
             1, 3.02, 4.08,
             1, 2.94, 3.78),
  sigma2 = 2,
  correlation = corAR1(value = 0.6, form = ~ hour|subject)
)

pwr.anova(design.rep)

}
