% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mkRTrms}
\alias{mkRTrms}
\title{Residual Variance-Covariance Matrices}
\usage{
mkRTrms(data, correlation = NULL)
}
\arguments{
\item{data}{a data frame with grouping factors and covariates.}

\item{corcall}{a \code{call} object specifying the residual correlation structure.
If NULL, an identity matrix is assumed.}
}
\value{
A list containing:
\itemize{
\item corcall
\item corframe: A processed data frame with indexed grouping variables and
ordering for correlation structures.
\item R: A block-diagonal residual variance-covariance structure, not yet scaled
by the residual variance
}
}
\description{
Residual Variance-Covariance Matrices
}
\keyword{internal}
