% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theta.names}
\alias{theta.names}
\title{Naming theta
Naming the vector in the order of model specification and in the actual order used in the model}
\usage{
theta.names(data, formula)
}
\arguments{
\item{data}{data frame}

\item{formula}{model formula}
}
\description{
Naming theta
Naming the vector in the order of model specification and in the actual order used in the model
}
