% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc.theta}
\alias{calc.theta}
\title{Calculate variance covariance parameters}
\usage{
calc.theta(VarCov, sigma)
}
\arguments{
\item{VarCov}{variance-covariance matrices. If there are multiple random effect groups,
supply the variance-covariance matrix of each group as an element in a list.}

\item{sigma}{standard deviation of random errors.}
}
\value{
theta
}
\description{
Scale variance-covariance matrices as the relative Cholesky factors of each random effect term.
}
\seealso{
"theta" in \code{\link[lme4]{getME}}, \code{\link[lme4]{lmer}}
}
