% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corClasses.R
\name{corClasses}
\alias{corClasses}
\alias{corAR1}
\alias{corARMA}
\alias{corCAR1}
\alias{corCompSymm}
\alias{corExp}
\alias{corGaus}
\alias{corLin}
\alias{corRatio}
\alias{corSpher}
\alias{corSymm}
\title{Correlation Structure Classes}
\usage{
corAR1(value = 0, form = ~1, fixed = FALSE)

corARMA(value = double(p + q), form = ~1, p = 0, q = 0, fixed = FALSE)

corCAR1(value = 0.2, form = ~1, fixed = FALSE)

corCompSymm(value = 0, form = ~1, fixed = FALSE)

corExp(
  value = numeric(0),
  form = ~1,
  nugget = FALSE,
  metric = c("euclidean", "maximum", "manhattan"),
  fixed = FALSE
)

corGaus(
  value = numeric(0),
  form = ~1,
  nugget = FALSE,
  metric = c("euclidean", "maximum", "manhattan"),
  fixed = FALSE
)

corLin(
  value = numeric(0),
  form = ~1,
  nugget = FALSE,
  metric = c("euclidean", "maximum", "manhattan"),
  fixed = FALSE
)

corRatio(
  value = numeric(0),
  form = ~1,
  nugget = FALSE,
  metric = c("euclidean", "maximum", "manhattan"),
  fixed = FALSE
)

corSpher(
  value = numeric(0),
  form = ~1,
  nugget = FALSE,
  metric = c("euclidean", "maximum", "manhattan"),
  fixed = FALSE
)

corSymm(value = numeric(0), form = ~1, fixed = FALSE)
}
\description{
These functions re-export \code{\link[nlme]{corClasses}} constructors from the \pkg{nlme} package.
}
