% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_calculator.R
\name{designCRD}
\alias{designCRD}
\title{#' Create Completely Randomized Design}
\usage{
designCRD(treatments, replications, formula, beta, sigma2, ...)
}
\arguments{
\item{treatments}{a vector representing the number of levels of treatment factors sequentially. For example, \code{treatments=c(2,3)} specifies two treatment factors with 2 and 3 levels, respectively.}

\item{replications}{number of experimental units in each group}

\item{formula}{an object of class \code{\link{formula}}: a symbolic description of the model used to assess treatment effects, which should contain fixed components and sources of variation according to experimental designs, assumptions, and objectives. The details of model specification for fixed effects see \code{\link{lm}} and random effects see \code{\link{lmer}}.}

\item{beta}{effect size: a vector including the expectations of intercept, main effects of factors and interactions between factors. Vector values are in the same order as the coefficients of the model that would be fitted using the exact data with \code{\link{contr.treatment}}.}

\item{sigma2}{error variance}

\item{...}{additional arguments to be passed to \code{fit.pseu.model}}
}
\value{
a list object containing design name, design data frame, model formula, and a pseudo model object with the expected effect size and variance-covariance components of random effects.
}
\description{
#' Create Completely Randomized Design
}
\examples{
mydesign <- designCRD(treatments = c(2, 2),
                      replications = 4,
                      beta = c(10, 9, 8, 7),
                      sigma2 = 10)
}
