% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_design.R
\name{df.crd}
\alias{df.crd}
\title{Create data frame for completely randomized design}
\usage{
df.crd(treatments, replications)
}
\arguments{
\item{treatments}{a vector representing the number of levels of treatment factors sequentially. For example, \code{treatments=c(2,3)} specifies two treatment factors with 2 and 3 levels, respectively.}

\item{replications}{number of experimental units in each group}
}
\value{
a data.frame with columns for treatment factors and replications
}
\description{
Create data frame for completely randomized design
}
\examples{
df.crd(5, 3)
}
