% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.R
\name{find_sample_size}
\alias{find_sample_size}
\title{Determine the sample size required to achieve target power}
\usage{
find_sample_size(
  design.quote,
  alpha = 0.05,
  target.power = 0.8,
  n_init = 2,
  n_max = 99,
  ...
)
}
\arguments{
\item{design.quote}{a quoted design object with replications that are unknown and evaluated,
denoted as "n"}

\item{alpha}{significance level, default 0.05}

\item{target.power}{The target power to achieve, default 0.8}

\item{n_init}{The initial sample size to start with, default 2}

\item{n_max}{The maximum sample size to try, default 99}

\item{...}{Additional arguments to pass to pwr.anova}
}
\value{
A data frame with the alpha, power, and best sample size for each sample size tried
}
\description{
This function finds the sample size required to achieve a target power for a given.
}
