% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit.pseu.model}
\alias{fit.pseu.model}
\title{Create an artificial model object}
\usage{
fit.pseu.model(formula, data, beta, VarCov, sigma, ...)
}
\arguments{
\item{formula}{an object of class \code{\link{formula}}: a symbolic description of the model used to assess treatment effects, which should contain fixed components and sources of variation according to experimental designs, assumptions, and objectives. The details of model specification for fixed effects see \code{\link{lm}} and random effects see \code{\link{lmer}}.}

\item{data}{a data frame with the independent variables of a design, e.g., treatment factors, block factors.}

\item{beta}{effect size: a vector including the expectations of intercept, main effects of factors and interactions between factors. Vector values are in the same order as the coefficients of the model that would be fitted using the exact data with \code{\link{contr.treatment}}.}

\item{VarCov}{variance-covariance components of random effects. If there are multiple random effects of one grouping factor, supply variance-covariance components in a matrix with the same order as the random effects specified in model formula. If there are multiple grouping factors, supply variance-covariance matrix for each grouping factor in a list.}

\item{sigma}{standard deviation of error}

\item{...}{other arguments passed to \code{\link{anova}}.}
}
\value{
a pseudo model object \code{lmerMod} or \code{\link{lm}}.
}
\description{
Create a pseudo model object with the same structure as the model object fitted using the exact data. Model coefficients are replaced by the effect size specified in the argument \code{beta}. The variance-covariance matrix of random effects is replaced by the argument \code{VarCov}. The standard deviation of error is replaced by the argument \code{sigma}.
}
