% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pwr.anova}
\alias{pwr.anova}
\title{Power calculation for main effects and interactions}
\usage{
pwr.anova(design, alpha, ...)
}
\arguments{
\item{design}{a list object containing design name, design data frame, model formula, and a pseudo model object with the expected effect size and variance-covariance components of random effects.}

\item{alpha}{significance level, type I error. default 0.05}

\item{...}{other arguments passed to anova. By default, type III sum of squares and Satterthwaite’s approximation of denominator degrees of freedom are applied.}
}
\value{
a data frame
}
\description{
Calculate power for testing overall effects of factors and interactions in the form of ANOVA table.
}
\examples{
mydesign = designCRD(treatments = c(2, 2),
                     beta = c(10, 9, 8, 7),
                     sigma2 = 9,
                     replications = 10)
pwr.anova(mydesign)

mydesign = designRCB(treatments = c(2, 2),
                      blocks = 10,
                      beta = c(10, 9, 8, 7),
                      VarCov = 8,
                      sigma2 = 9)
pwr.anova(mydesign)
}
