% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-calculation.R
\name{pwr.anova}
\alias{pwr.anova}
\title{Power calculation for omnibus test}
\usage{
pwr.anova(design, alpha, ...)
}
\arguments{
\item{design}{a design object created using design generating functions, \code{\link{designCRD}}, \code{\link{designRCBD}},
\code{\link{designLSD}},  \code{\link{designCOD}}, \code{\link{designSPD}}, and \code{\link{designCustom}}}

\item{alpha}{significance level (type I error rate), default 0.05}

\item{...}{#' Other arguments are passed to the \code{anova} function in \code{lmerTest} for linear mixed models
and to the \code{Anova} function in \code{car} for linear models.
The type of ANOVA table, with Type III as the default, and the method for computing the denominator degrees
of freedom, with Satterthwaite's method as the default, can be changed.
For more details, see \link[lmerTest]{anova.lmerModLmerTest} and \link[car]{Anova}.
For linear mixed models, these arguments should also be defined in the design-generating functions
before passing the design object to the \code{pwr.anova} function.}
}
\value{
a data frame
}
\description{
Calculate power for testing overall effects of treatment factors and their interactions, i.e., ANOVA test.
}
\examples{
# generate an RCBD
rcbd = designRCBD(treatments = c(2, 2), blocks = 10, beta = c(10, 9, 8, 7), VarCov = 10, sigma2 = 9)
pwr.anova(rcbd)
}
