% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pwr.contrast}
\alias{pwr.contrast}
\title{Power calculation for contrasts}
\usage{
pwr.contrast(design, specs, contrast, alpha = 0.05, ...)
}
\arguments{
\item{design}{a list object containing design name, design data frame, model formula, and a pseudo model object with the expected effect size and variance-covariance components of random effects.}

\item{specs}{see \code{\link[emmeans]{emmeans}}}

\item{contrast}{see \code{\link[emmeans]{contrast}}}

\item{alpha}{significance level, type I error. default 0.05}

\item{...}{other arguments passed to \code{\link[emmeans]{contrast}}. By default, kenward-roger approximation for degrees of freedom is applied.}
}
\value{
a data frame
}
\description{
Calculate power for testing various contrasts of fixed effects. The same syntax of \code{\link{emmeans}} package is employed to specify contrast types.
}
\examples{
mydesign = designRCB(treatments = c(2, 2),
                     blocks = 10,
                     beta = c(10, 9, 8, 7),
                     VarCov = 8, sigma2 = 9)
pwr.contrast(mydesign, ~ facA|facB, contrast = "pairwise")
}
