% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-calculation.R
\name{pwr.summary}
\alias{pwr.summary}
\title{Power of model coefficients}
\usage{
pwr.summary(object, sig.level = 0.05)
}
\arguments{
\item{object}{a design object created in pwr4exp}

\item{sig.level}{significance level, default 0.05}
}
\value{
The power for testing model coefficients
}
\description{
Power of model coefficients
}
\examples{
rcbd <- designRCBD(
  treatments = c(2, 2),
  label = list(facA = c("1", "2"), facB = c("1", "2")),
  blocks = 12,
  formula = ~ facA*facB + (1|block),
  means = c(32, 35, 30, 37),
  vcomp = 4,
  sigma2 = 6
)
pwr.summary(rcbd)
}
